+++ travis/install-redis.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+sudo apt-get install redis-server
+sudo service redis-server start